split edit mode and read-only
